+++ test/highlight_text_plus_test.dart
@@ -1,7 +1,3 @@
-import 'package:flutter_test/flutter_test.dart';
-
-import 'package:highlight_text_plus/highlight_text_plus.dart';
-
 void main() {
 
 }